restructured, refactored
