--- json_cfg/__init__.py
@@ -1,1 +1,0 @@
-from .holder import *
