--- logger/__init__.py
@@ -1,1 +1,0 @@
-from .log_data import LogsWriter
