--- logger/filenames.py
@@ -1,3 +1,0 @@
-EXCEPTIONS = "files/logs/errors.txt"
-RECEIVED_MESSAGES = "files/logs/received_messages.txt"
-SENT_MESSAGES = "files/logs/sent_messages.txt"
