+++ main.py
@@ -8,3 +8,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+# TODO: add music downloading
