+++ message/__init__.py
@@ -1,0 +1,5 @@
+from .attachments.doc import Doc
+from .attachments.photo import Photo
+from .attachments.voice_message import Voice
+
+from .message.message import TextMessage
