+++ structs/__init__.py
@@ -1,4 +1,7 @@
-from .event_structs import *
-from .new_message import *
-from .attachment_types import *
-from .launch_config import *
+from .config.db import *
+from .config.launch import *
+from .config.messages import *
+
+from .events.exceptions import *
+from .events.received_message import *
+from .events.sent_message import *
